+++ src/cmd/mod.rs
@@ -84,7 +84,7 @@ trait Cmd {
 }
 
 struct GlobalArgs {
-    quiet: bool,
+    _quiet: bool,
     _verbose: bool,
     no_styles: bool,
 }
@@ -92,7 +92,7 @@ struct GlobalArgs {
 impl GlobalArgs {
     fn new(quiet: bool, verbose: bool, no_styles: bool) -> Self {
         Self {
-            quiet,
+            _quiet: quiet,
             _verbose: verbose,
             no_styles,
         }